dm-3111
move helper methods for faciity index page for CRHs to helper mod
